nvshmem_opt: Add custom barrier for jacobi
This commit adds a custom barrier for the jacobi app.
The jacobi app in this repo does not need a global
barrier. Rather, each rank just needs to synchronize
with its neighbor before moving on to the next iteration.
This custom barrier does exaclty that and hence will
be faster than using shmem_barrier_all, especially
when we strong scale to large PE count.